requirements.txt
# requirements.txt
PyQt5>=5.15.0
yt-dlp>=2024.1.1
ffmpeg-python>=0.2.0
pyinstaller>=5.13.0
python-dotenv==1.0.1
cryptography>=41.0.7
certifi>=2024.2.2
requests>=2.31.0
typing-extensions==4.12.2
packaging==23.1
browser-cookie3>=0.19.1
pathlib>=1.0.1